requirements.txt
PyYAML
requests
urllib3==1.26.15